+++ lib/widgets/search/entry_type_filter.dart
@@ -19,7 +19,6 @@ class EntryTypeFilter extends StatelessWidget {
           child: WrapSuper(
             alignment: WrapSuperAlignment.center,
             spacing: 5,
-            lineSpacing: 5,
             children: [
               ...entryTypes.map(EntryTypeChip.new),
               const EntryTypeAllChip(),