+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed:
 - Navigate back icon on task page
 
+### Changed:
+- Improved task input fields layout
+
 ## [0.8.264] - 2023-02-11
 ### Fixed:
 - Audio playback restart on navigate to audio entry