+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed:
 - Improved task input fields layout
+- Improved chip layout
 
 ## [0.8.264] - 2023-02-11
 ### Fixed: